# Optimize Service Stats