… common/appscale/common/service_stats/categorizers.py
… common/appscale/common/service_stats/matchers.py
… common/appscale/common/service_stats/metrics.py
… common/test/unit/test_stats_service/test_categorizers.py
… common/test/unit/test_stats_service/test_matchers.py
… common/test/unit/test_stats_service/test_metrics.py
… common/test/unit/test_stats_service/test_stats_manager.py