All these changes were made for Service Stats' performance improvement.

New version uses functions instead of classes for implementing categorizers, matchers and metrics.
Converts configurations dict to configurations tuple of tuples.
Restores stats collection from tq servers.
Modifies old tests.
Deletes unused modules.

Furthermore, now user should do initialization of all request's fields on their own, not to pass them as arguments to `start_request()` and `finalize()` methods. It also makes Stats Service work faster.