@Honcharov12 2019-01-24
Demo:

Were generated 100000 request by following rule:
```
for i in range(100000):
  if i % 2 == 0:
    app = 'guestbook'
  elif i % 3 == 0:
    app = 'ghostbook'
  else:
    app = 'other'
  
  if i % 4 == 0:
    method = 'GET'
  else:
    method = 'POST'

  if i % 5 == 0:
    resource = '/'
  elif i % 13 == 0:
    resouce = '/tmp'
  else:
    resource = '/home/user'

  if i % 7 == 0:
    status = 400 + i % 3
  elif i % 11 == 0:
    status = 500 + i % 7
  else:
    status = 200 + i % 5
  requests_info.append((app, method, resource, status))
```

Then they were processed.
In current version of Service Stats:
```
class FailedMatcher(matchers.RequestMatcher):
  def matches(self, request_info):
    return 400 <= request_info.status <= 599
FAILED = FailedMatcher()

cumulative_dict  = {
  'all': matchers.ANY,
  '4xx': matchers.CLIENT_ERROR,
  '5xx': matchers.SERVER_ERROR,
  'latency': summarizers.AttrValue('latency'),
  'failed': FAILED,
  categorizers.ExactValueCategorizer('by_app', field='app'): {
    'all': matchers.ANY,
    '4xx': matchers.CLIENT_ERROR,
    '5xx': matchers.SERVER_ERROR,
    'latency': summarizers.AttrValue('latency'),
    'failed': FAILED,
    categorizers.ExactValueCategorizer('by_method', field='method'): {
      'all': matchers.ANY,
      'failed': FAILED,
      'latency': summarizers.AttrValue('latency')
    }
  }
}
stats = stats_manager.ServiceStats('my_service', cumulative_counters=cumulative_dict)

for app, method, resource, status in requests_info:
  req = stats.start_request(app=app, method=method, resource=resource)
  stats.get_recent() # were tests with and without it
  req.finalize(status=status)
```
Without getting recent statistics all requests were processed in 1.365 secs, and with - in 44.714 secs.

After modification
```
FAILED = lambda req_info: 400 <= req_info.status <= 599

cumulative_dict = {
  'all': samples.summarize_any,
  '4xx': samples.summarize_client_error,
  '5xx': samples.summarize_server_error,
  'latency': samples.summarize_latency,
  'failed': FAILED, 
  ('by_app', samples.categorize_by_app): {
    'all': samples.summarize_any,
    '4xx': samples.summarize_client_error,
    '5xx': samples.summarize_server_error,
    'latency': samples.summarize_latency,
    'failed': FAILED,
    ('by_method', samples.categorize_by_method): {
      'all': samples.summarize_any,
      'failed': FAILED,
      'latency': samples.summarize_latency
    }
  }
}

stats = stats_manager.ServiceStats('my_service', cumulative_counters=cumulative_dict)
for app, method, resource, status in requests_info:
  req = stats.start_request()
  req.app = app
  req.method = method
  req.resource = resource
  stats.get_recent()
  req.status = status
  req.finalize()
```
were such results: 
 - without getting recent statistics - 0.561 secs;
 - with getting recent statistics - 12.616 secs.